// projects/workflow-multi/src/lib/interface/diagram-action-type.enum.ts
export enum DiagramActionType {
  CLICK_ON_GRID = 'CLICK_ON_GRID',
  CONTEXT_MENU_ON_STEP = 'CONTEXT_MENU_ON_STEP',
  CONTEXT_MENU_ON_CONNECTOR = 'CONTEXT_MENU_ON_CONNECTOR',
  ADD_STEP = 'ADD_STEP',
  ADD_CONNECTOR = 'ADD_CONNECTOR',
  UPDATE_CONNECTOR = 'UPDATE_CONNECTOR',
  CLICK_ON_ELEMENT = 'CLICK_ON_ELEMENT',
  CONTEXT_MENU_ON_GRID = 'CONTEXT_MENU_ON_GRID',
  ZOOM = 'ZOOM',
}